Rename Backtrace::InspectResult.inspect to .inspect_result
The current implementation of `Backtrace::InspectResult.inspect` accepts an object to interpret and stringify, but calling `inspect` on an object should returna human-readable string of the invoked object, not its arguments. Ruby encourages the overriding of this method, but `Backtrace::InspectResult.inspect` was never intended to replace the default definition of `inspect` since it provides new functionality with a different arity.