Fix edge case in block string lexer
When there are 2 or more quote characters within a block string, and
there is a single non-quote character between them, the lexer failed
to match a block quote.

Fixing this also results in a simpler machine. The incorrectness of the
previous machine resulted in an extra case that can now be removed.