Allow code to be accessed outside of the errors
This allows better reporting of codes without the need to use `to_h`. It's also better for subclasses as they can modify if they need to. Which for fine grained reporting they would want to.

I'm sure I had good reasons to make this private originally, but use cases now show this would be a much better solution.