Fix handling of invalid UTF-8 bytes
If the string isn't even validly encoded bytes (let alone invalid
escapes, which is already being checked) then the regex itself raises an
ArgumentError which bubbles all the way up and leaves everything in a
bad state.

I had to put the string in a separate file because Ruby completely
refuses to run the test file if it contains a string literal with
encoding this messed up.