Fix argument default value (Ruby Date) not being serialized
This commit addresses an issue whereby using a Ruby Date object as the
default value for an argument was being serialized to `nil` instead of
either serializing to a `Date` or raising an error. The fix contained in
this commit serializes the object as a `Date`.

Fixes https://github.com/rmosolgo/graphql-ruby/issues/3520