Ensure error size is linear in the number of conflicts
Currently error size is exponential in the size of conflicts - every
combination is reported. For example for 2 conflicts, we get the
expected `a or b`, but for three we get: `(a or b), or (a or c) or (b or
c)`. This scales poorly, and can result in performance issues on
maliciously crafted input.

This change ensures that error reporting is linear in the size of the
number of conflicts.